Main page setengah jalan, tinggal edit create dan component per card
